Move LLVMOptions to an options package
