+++ projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/LLVMLogger.java
@@ -30,6 +30,7 @@
 package com.oracle.truffle.llvm.runtime;
 
 import com.oracle.truffle.api.CompilerAsserts;
+import com.oracle.truffle.llvm.runtime.options.LLVMOptions;
 
 public class LLVMLogger {
