+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/RemoteTestSuiteBase.java
@@ -46,7 +46,7 @@ import org.junit.BeforeClass;
 
 import com.oracle.truffle.llvm.LLVM;
 import com.oracle.truffle.llvm.runtime.LLVMLogger;
-import com.oracle.truffle.llvm.runtime.LLVMOptions;
+import com.oracle.truffle.llvm.runtime.options.LLVMOptions;
 import com.oracle.truffle.llvm.tools.util.ProcessUtil;
 
 public class RemoteTestSuiteBase extends TestSuiteBase {
