+++ projects/com.oracle.truffle.llvm.tools/src/com/oracle/truffle/llvm/tools/LLVMToolPaths.java
@@ -31,7 +31,7 @@ package com.oracle.truffle.llvm.tools;
 
 import java.io.File;
 
-import com.oracle.truffle.llvm.runtime.LLVMOptions;
+import com.oracle.truffle.llvm.runtime.options.LLVMOptions;
 
 public class LLVMToolPaths {
